fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16389 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSLane.cpp
@@ -1290,7 +1290,7 @@ int
 MSLane::getCrossingIndex() const {
     for (MSLinkCont::const_iterator i = myLinks.begin(); i != myLinks.end(); ++i) {
         if ((*i)->getLane()->getEdge().isCrossing()) {
-            return i - myLinks.begin();
+            return (int)(i - myLinks.begin());
         }
     }
     return -1;
